+++ Delight/Utility/DLRecordingContext.h
@@ -21,7 +21,6 @@ typedef enum {
 @interface DLRecordingContext : NSObject <NSCoding>
 
 @property (nonatomic, retain) NSString * sessionID;
-@property (nonatomic, retain) NSString * appUserID;
 @property (nonatomic, retain) NSDictionary * tracks;
 @property (nonatomic, retain) NSMutableDictionary * sourceFilePaths;
 @property (nonatomic) BOOL shouldRecordVideo;
@@ -34,6 +33,7 @@ typedef enum {
 @property (nonatomic, retain) NSString * cameraFilePath;
 @property (nonatomic, retain) NSString * touchFilePath;
 @property (nonatomic, retain) NSString * usabilityTestDescription;
+@property (nonatomic, retain) NSMutableDictionary * userProperties;
 @property (nonatomic, retain) NSMutableIndexSet * finishedTaskIndex;
 @property (nonatomic) BOOL saved;
 @property (nonatomic, readonly) BOOL loadedFromArchive;